Makefile.kdevops: enable disabling force from installing ansible deps
I'm not a big fan of installing ansible-galaxy roles with "--force",
even though without it you won't get updates... this seems to be a
"feature" of ansible, not an issue. Well, I hate that decision.

So you might hate it too, and so enable you to disable force when you
know you don't need it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>