upgrade devconfig and terraform roles
The terraform role has updates to update gce to our current
definition of what we want to kdevops, and the provider. In
particular this enables ansible, and adds a new scratch disk.

The devconfig update ensures we don't run into snags when libraries
are updated which would otherwise update packages but not ssh. We
address this by doing the full distro upgrade at once.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>