update the ssh update python scripts to use python3 statically
We use a python script to update your localhost ssh .config based on
the hosts provisioned. For vagrant we use an ansible role for that,
and for terraform we use a terraform module for that. When we update
the terraform module we must update the terraform ansible role for it.

We update both scripts to statically rely on python3. We do this
as some distributions don't want you to use /usr/bin/env and also
want you to use the actual version of python you want. For example
the Debian Python Policy:

https://www.debian.org/doc/packaging-manuals/python-policy/ch-python.html#s-interpreter

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>