kdevops: update roles to support a custom ssh kexalgorithms
The python script has been updated significantly to support
custom ssh kexalgorithms, which is required for old hosts
such as SLES10-SP3.

We enable support for this on vagrant and and on terraform.
kdevops will soon have support for this via kconfig for
vagrant hosts, however I'm not sure if we'll ever have
a host on the cloud needing this. We'll see!

The vagrant ansible role also has been modified to just run
locally. This means vagrant's playbook list should be updated
to remove the update ssh role, and instead the controller
should call that manually throught the Makefile. This is
done because:

1) There is no need to run ansible on a target host to update
   its ssh configuration locally.
2) Even if we tried the above we still were asking the tasks
   to run locally anyway
3) By doing this we can support updating your ssh configuration
   for even hosts which do not support the latest ansible, which
   require Python 2.6. For instance, SLES10-SP3 has Python 2.4.

This should in the end enable us to add the ssh entries for
SLES10-SP3 on to your ssh configuration after provisioning
with vagrant.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>